aas-sim/backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aas-sim-backend"
version = "0.1.0"
description = "AAS Simulation Backend"
authors = [{name = "Developer", email = "dev@example.com"}]
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "python-multipart>=0.0.6",
    "pyyaml",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "httpx>=0.25.2",
]

[tool.hatch.envs.default]
dependencies = [
    "pytest",
    "pytest-asyncio",
    "httpx",
]

[tool.hatch.envs.default.scripts]
test = "pytest {args:tests}"
dev = "uvicorn app.main:app --reload --host 0.0.0.0 --port 8000"